chore: fix route instantiation, add nesting
