+++ web-registry/src/App.tsx
@@ -137,49 +137,15 @@ const App: React.FC = (): JSX.Element => {
               element={<ProtectedRoute component={OrganizationProfile} />}
             />
             <Route
-              path="project-pages"
+              path="project-list"
               element={<ProtectedRoute component={ProjectList} />}
             />
-            <Route path="project-pages/:projectId" element={ProjectCreate}>
-              <Route
-                path="choose-credit-class"
-                element={<KeplrRoute component={ChooseCreditClassPage} />}
-              />
-              <Route
-                path="basic-info"
-                element={<KeplrRoute component={BasicInfo} />}
-              />
-              <Route
-                path="location"
-                element={<KeplrRoute component={ProjectLocation} />}
-              />
-              <Route path="story" element={<KeplrRoute component={Story} />} />
-              <Route
-                path="description"
-                element={<KeplrRoute component={Description} />}
-              />
-              <Route path="media" element={<KeplrRoute component={Media} />} />
-              <Route
-                path="metadata"
-                element={<KeplrRoute component={ProjectMetadata} />}
-              />
-              <Route path="roles" element={<KeplrRoute component={Roles} />} />
-              <Route
-                path="entity-display"
-                element={<KeplrRoute component={EntityDisplay} />}
-              />
-              <Route
-                path="review"
-                element={<KeplrRoute component={ProjectReview} />}
-              />
-              <Route
-                path="finished"
-                element={<KeplrRoute component={ProjectFinished} />}
-              />
-              <Route
-                path="edit"
-                element={<KeplrRoute component={ProjectEdit} />}
-              >
+            <Route path="project-pages">
+              <Route path=":projectId" element={<ProjectCreate />}>
+                <Route
+                  path="choose-credit-class"
+                  element={<KeplrRoute component={ChooseCreditClassPage} />}
+                />
                 <Route
                   path="basic-info"
                   element={<KeplrRoute component={BasicInfo} />}
@@ -193,8 +159,16 @@ const App: React.FC = (): JSX.Element => {
                   element={<KeplrRoute component={Story} />}
                 />
                 <Route
+                  path="description"
+                  element={<KeplrRoute component={Description} />}
+                />
+                <Route
                   path="media"
                   element={<KeplrRoute component={Media} />}
+                />
+                <Route
+                  path="metadata"
+                  element={<KeplrRoute component={ProjectMetadata} />}
                 />
                 <Route
                   path="roles"
@@ -204,6 +178,43 @@ const App: React.FC = (): JSX.Element => {
                   path="entity-display"
                   element={<KeplrRoute component={EntityDisplay} />}
                 />
+                <Route
+                  path="review"
+                  element={<KeplrRoute component={ProjectReview} />}
+                />
+                <Route
+                  path="finished"
+                  element={<KeplrRoute component={ProjectFinished} />}
+                />
+                <Route
+                  path="edit"
+                  element={<KeplrRoute component={ProjectEdit} />}
+                >
+                  <Route
+                    path="basic-info"
+                    element={<KeplrRoute component={BasicInfo} />}
+                  />
+                  <Route
+                    path="location"
+                    element={<KeplrRoute component={ProjectLocation} />}
+                  />
+                  <Route
+                    path="story"
+                    element={<KeplrRoute component={Story} />}
+                  />
+                  <Route
+                    path="media"
+                    element={<KeplrRoute component={Media} />}
+                  />
+                  <Route
+                    path="roles"
+                    element={<KeplrRoute component={Roles} />}
+                  />
+                  <Route
+                    path="entity-display"
+                    element={<KeplrRoute component={EntityDisplay} />}
+                  />
+                </Route>
               </Route>
             </Route>
             <Route path="admin" element={<Admin />} />
