+++ web-registry/src/pages/ProjectCreate/index.ts
@@ -1,3 +1,3 @@
 export { ProjectCreate as default } from './ProjectCreate';
 
-export { useCreateProjectContext } from './ProjectCreate.context';
+export { useCreateProjectContext } from './ProjectCreate';
